feat: find files to rename to pkg
