+++ api/communication/communication.proto
@@ -1,7 +1,7 @@
 
 syntax = "proto3";
 
-option go_package = "github.com/Ow1Dev/noctifunc/pkgs/api/communication";
+option go_package = "github.com/Ow1Dev/noctifunc/pkg/api/communication";
 
 service CommunicationService {
   rpc Execute(ExecuteRequest) returns (ExecuteResponse);
